+++ CHANGELOG.md
@@ -4,7 +4,37 @@ All notable changes to Vox are documented here.
 
 ## [Unreleased]
 
-## [0.10.0-rc.0] - 2026-06-18
+## [0.10.0-rc.6](https://github.com/facet-rs/facet/compare/vox-codegen-v0.10.0-rc.5...vox-codegen-v0.10.0-rc.6) - 2026-06-28
+
+### Other
+
+- updated the following local packages: facet-core, facet, vox-phon, vox-types, phon-schema, phon-ir, phon-engine, phon, phon-codegen
+
+## [0.10.0-rc.5](https://github.com/facet-rs/facet/compare/vox-codegen-v0.10.0-rc.4...vox-codegen-v0.10.0-rc.5) - 2026-06-26
+
+### Other
+
+- updated the following local packages: facet-core, facet, vox-phon, vox-types, phon-schema, phon-ir, phon-engine, phon, phon-codegen
+
+## [0.10.0-rc.4](https://github.com/facet-rs/facet/compare/vox-codegen-v0.10.0-rc.3...vox-codegen-v0.10.0-rc.4) - 2026-06-26
+
+### Other
+
+- updated the following local packages: facet-core, facet, phon-schema, phon-ir, phon-engine, phon, vox-phon, vox-types, phon-codegen
+
+## [0.10.0-rc.3](https://github.com/facet-rs/facet/compare/vox-codegen-v0.10.0-rc.2...vox-codegen-v0.10.0-rc.3) - 2026-06-25
+
+### Other
+
+- updated the following local packages: facet-core, facet, phon-schema, phon-ir, phon-engine, phon, vox-phon, vox-types, phon-codegen
+
+## [0.10.0-rc.2](https://github.com/facet-rs/facet/compare/vox-codegen-v0.10.0-rc.1...vox-codegen-v0.10.0-rc.2) - 2026-06-22
+
+### Other
+
+- Fix Swift codegen trailing whitespace
+
+## [0.10.0-rc.1] - 2026-06-18
 
 ### Other